Updated somes domain class.

+++ lib/features/recipe/domain/entities/ingredient.dart
@@ -2,7 +2,7 @@ import 'dart:ffi';
 
 class Ingredient {
   String name;
-  Float quantity;
+  double quantity;
   String measure; //TODO:enum or new class
 
   Ingredient(this.name, this.quantity, this.measure);
